this should all be in genetic maps
