--- forrustts-genetic-maps-rand/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "forrustts-genetic-maps-rand"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rand = "0.8.5"
-rand_distr = "0.4.3"
-forrustts-genetic-maps = {version = "0.1.0", path="../forrustts-genetic-maps"}
-forrustts-core = {version="0.1.0", path="../forrustts-core"}
